docs(operators): add documentation for bufferWhen (#236)
* docs(operations): add documentation for bufferWhen to close #113

Add documentation for operator bufferWhen to close #113

#113

* fix(build): Fix Travis CI build

Add dependency for @angular-devkit/core to fix the Travis CI build

* revert(dependencies): Modify package.json to use the latest angular-cli

As an temporary solution I added @angular-devkit/core as a devDependency to solve the build issues
from Travis. However, I now changed the package.json to include the latest angular-cli version (to
1.6.5) and this allows Travis to build correctly. I find this a better solution than the previous
one

* fix(operators): Modify windowTime to windowWhen

The related operators was listing windowTime, but instead should be listing windowWhen

* refactor(operators): Add comment for expected output on console

Added comment in the documentation as requested to contain the expected console output. Removed the
$ suffix from the examples for now for consistency (wait for coding guidelines).

* style(operators): Move expected console output after subscription

The expected console output should be after the subscription. Also deleted the generics from the
method signature.