feat($browser): Added support for Server Side Rendering (Angular Universal)
Added support for Server Side Rendering (Angular Universal). The major difference is that now, upon
build, the browser code goes to dist/browser and the server code goes to dist/server, respectively.
However, its up to the backend team to implement SSR on the server.

The only breaking change is the directory structure: browser code used to go in /dist, now there are
two folders - /dist/browser and /dist/server