feat(operators): Add signature based on parameters collection
Resolve signature from name, parameters and return type values as per #168. This change includes an
update to the OperatorDoc type, because the return type of every operator needed to be included. All
existing operators have been updated as well to include the correct return type. Some updates to the
css were required to add colors to the signature. Added a new pipe (ArgumentPipe) to create a nice
tooltip when hovering over the arguments.